refactor(alerts): extract natural language draft compiler
